+++ lib/logger/test/test_helper.exs
@@ -1,3 +1,4 @@
+Logger.configure_backend(:console, colors: [enabled: false])
 ExUnit.start()
 
 defmodule Logger.Case do